+++ CHANGES.rst
@@ -8,6 +8,8 @@ Unreleased
 
 - 301 redirects result in a ``Redirect`` exception.
 - ``Requestor`` is now initialzed with a ``timeout`` parameter.
+- ``ScriptAuthorizer``, ``ReadOnlyAuthorizer``, and ``DeviceIDAuthorizer`` have a
+  new parameter, ``scopes``, which determines the scope of access requests.
 
 2.2.1 (2021/07/06)
 ------------------